backend/requirements.txt
# Direct dependencies only — run `pip install -r requirements.txt`
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-jose[cryptography]==3.5.0
bcrypt==4.1.3
pydantic[email]==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.22
starlette==0.37.2

# Dev / testing
requests==2.32.5
pytest==9.0.2
